introduce cmake as build system (proof of concept)
Please review the contents of the CMakeLists.txt files;
they are pretty easy to understand and to extend.

I chose cmake as libsmatrix (as a embedded external project)
is very easy to integrate into dawanda_recommender project.

I considered using automake, too, but cmake was just easier to get
running. Feel free to counter this with an automake-solution,
that easily integrates into dawanda_recommender, too. ;-)